chore: move globals store into specs store as well
